* Fixed compiler warnings on windows

+++ libsrc/util/uuid.c
@@ -96,15 +96,15 @@ uuid_unpack (const uuid_t in, struct uuid *uu)
 
   tmp = *ptr++;
   tmp = (tmp << 8) | *ptr++;
-  uu->time_mid = tmp;
+  uu->time_mid = (uint16) tmp;
 
   tmp = *ptr++;
   tmp = (tmp << 8) | *ptr++;
-  uu->time_hi_and_version = tmp;
+  uu->time_hi_and_version = (uint16) tmp;
 
   tmp = *ptr++;
   tmp = (tmp << 8) | *ptr++;
-  uu->clock_seq = tmp;
+  uu->clock_seq = (uint16) tmp;
 
   memcpy (uu->node, ptr, 6);
 }
@@ -139,16 +139,16 @@ uuid_parse (const char *in, uuid_t uu)
 	return -1;
     }
   uuid.time_low = strtoul (in, NULL, 16);
-  uuid.time_mid = strtoul (in + 9, NULL, 16);
-  uuid.time_hi_and_version = strtoul (in + 14, NULL, 16);
-  uuid.clock_seq = strtoul (in + 19, NULL, 16);
+  uuid.time_mid = (uint16) strtoul (in + 9, NULL, 16);
+  uuid.time_hi_and_version = (uint16) strtoul (in + 14, NULL, 16);
+  uuid.clock_seq = (uint16) strtoul (in + 19, NULL, 16);
   cp = in + 24;
   buf[2] = 0;
   for (i = 0; i < 6; i++)
     {
       buf[0] = *cp++;
       buf[1] = *cp++;
-      uuid.node[i] = strtoul (buf, NULL, 16);
+      uuid.node[i] = (uint8) strtoul (buf, NULL, 16);
     }
 
   uuid_pack (&uuid, uu);
